Added a fix for dirty annotations. Corrupted annotations are cleaned rigth before extracting metadata.
When deleting or inserting annotations, SystemPlus shifts many annotations in the file to keep all notes in ascending order of time. This seems not always to be done cleanly. Traces of older annotations may linger in memory slots of new annotations. SystemPlus is not bothered as it only seems to read the annotation up untill a null character.
Instead of rewriting all dirty annotations manually in SystemPlus, this fix allows to ignore the part after a null character.